% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{geostatsgpu}
\alias{geostatsgpu}
\alias{geostatsgpu-package}
\title{Geostats with GPU}
\description{
Matern correlation function computed in gpu.
}
\author{
Patrick Brown <patrick.brown@utoronto.ca>
}
