% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viennaclLu.R
\name{luGpu}
\alias{luGpu}
\title{LU decomposition}
\usage{
luGpu(x, D)
}
\arguments{
\item{x}{matrix}
}
\description{
LU decomposition using ViennaCL
}
\examples{
library('geostatsgpu')
gpuR::setContext(grep('gpu', listContexts()$device_type) [1])
v1 = matrix(1,10,10)
diag(v1) = 2
v2 <- gpuR::vclMatrix(v1, type='float')
v3 = gpuR::deepcopy(v2)
system.time(res1 <- luGpu(v2))
system.time(res2 <- chol(v1))
system.time(res3 <- cholGpu(v3))

}
