% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternVcl.R
\name{maternGpu}
\alias{maternGpu}
\title{Matern correlation}
\usage{
maternGpu(x, param = c(range = 1, variance = 1, shape = 1),
  output = vclMatrix(data = 0, nrow(x), nrow(x), type = "double", ctx_id
  = x@.context_index), type = c("variance", "cholesky", "precision",
  "inverseCholesky"), DofLDL = NULL)
}
\arguments{
\item{x}{points}

\item{param}{parameters}

\item{output}{VCL matrix to store result}

\item{type}{produce variance matrix or it's cholesky}

\item{DofLDL}{a VCL vector to hold the diagonals from the LDL decomposition}
}
\value{
The output and DofLDL objects have been altered in-place
}
\description{
Compute Matern covariance function
}
\examples{
x = as.matrix(expand.grid(seq(1,by=1,len=23), seq(201,by=1,len=14)))
coordsV = gpuR::vclMatrix(x)
D3 <- vclMatrix(
    data=-1, 
    nrow(coordsV), nrow(coordsV),
    type='double'
)

myParams = c(shape=4.5, range=1.5, variance = 2, nugget = 0, anisoRatio = 2, anisoAngleRadians = pi/4)

system.time(v1 <- maternGpu(coordsV, myParams, D3, 'variance'))
system.time(mmat <- geostatsp::matern(sp::SpatialPoints(as.matrix(coordsV)), myParams))
as.matrix(v1)[1:5,1:5]
mmat[1:5,1:5]

}
